Create BatchProcessor objects based on need.
Rather than creating a batch processor for each event type that is
received, create one for each event type that has a tap and destroy
those that no longer have taps. This ensures that events are not
needlessly queued, only to be dropped because there are no taps for
them.