Remove EventTapWriter from downstream of the BatchProcessor
The BatchCloner is now downstream of the BatchProcessor and EventTapFlow
is downstream of the former. This reduces the need for synchronization
in the EventTapWriter to only when updating the configuration based on
discovery.

In order to maintain information about a given event type, the
EventTypeInfo inner class was created. Updating the configuration will
create new EventTypeInfo objects but will only create new
BatchProcessor, EventTapFlow, and BatchCloner objects if absolutely
necessary.