Create HttpEventTapFlowFactory
Guice's FactoryModuleBuilder doesn't really support what is needed for
creating HttpEventTapFlow objects, which will need slightly different
parameters, depending on whether a QoS flow is being created or not.
Replace that with an explicit factory, which will do the right thing
when support for QoS is added to the flow.