Support prioritized event types.
Using 3 ScheduledExecutorService objects, support prioritized event
types. Those that are considered high priority, schedule combining in an
executor with a pool of threads large enough to have one for each type.
Those that are considered low priority, schedule combining in an
executor with a single thread. For all other event types, schedule in an
executor with a pool of threads of configurable size.