Work around maven dependency issue
For a currently unknown reason, the fact that aws-java-sdk dependency
for either 1.1.1 or 1.1.2 of commons-logging, in combination with other
modules depending explicitly on 1.1, causes maven to package 1.1.2 but
lists 1.1.1 in the manifest. Until the root cause can be found and
fixed, exclude commons-logging dependency from aws-java-sdk so that this
can be properly packaged.